Include "Client-Development" expenses in 'client'
This allows us to include spec work when looking at our expenses.

(Also, add FIXME comments about some related possible improvements.)

I did a rudimentary regression test for this change, which is
something that we should formalize for future changes.  I just ran
this before the change...

  $ for name in CLIENT1 CLIENT2 ...; do
    echo "*** ${name}: ***" >> 1-client-report.out
    echo "" >> 1-client-report.out
    ./qb client ${name} 2017-2022 >> 1-client-report.out
    echo "" >> 1-client-report-orig.out
  done

...and then this after the change...

  $ for name in CLIENT1 CLIENT2 ...; do
    echo "*** ${name}: ***" >> 2-client-report.out
    echo "" >> 2-client-report.out
    ./qb client ${name} 2017-2022 >> 2-client-report.out
    echo "" >> 2-client-report-orig.out
  done

...then made sure there were no differences (or only the expected
differences, but as it happens with the particular clients I used no
differences were to be expected):

  $ diff -u 1-client-report-orig.out 2-client-report-orig.out